Added variant file definition and features from jesup
Docs updated to reflect android/desktop capabilities
