+++ desktop.sh
@@ -1,0 +1,7 @@
+#!/bin/bash
+
+$BROWSERTIME_BIN \
+    --browser firefox \
+    --skipHar \
+    --firefox.binaryPath="$FIREFOX_BINARY_PATH" \
+    "$@"
