+++ preload_slim.js
@@ -1,0 +1,4 @@
+module.exports = async function(context, commands) {
+  let url = context.options.browsertime.url;
+  return commands.measure.start(url);
+};
