+++ sites.txt
@@ -1,3 +1,5 @@
+https://www.foxnews.com/sports/nfl-stars-lowes-donate-1-million-christmas-trees-families-need
+https://www.wsj.com/articles/trump-administration-claims-facebook-improperly-reserved-jobs-for-h1-b-workers-11607023044
 https://accounts.google.com
 https://cnn.com/ampstories/us/why-hurricane-michael-is-a-monster-unlike-any-other
 https://discordapp.com/
