+++ variants_desktop.py
@@ -1,0 +1,3 @@
+# Define the apps to test
+#  The last parameter can be a firefox pref string (e.g '--firefox.preference network.http.rcwn.enabled:false ')
+variants = [('nightly', 'desktop.sh', '' )]
